+++ CHANGELOG.rst
@@ -17,6 +17,9 @@
 - Detect and flag "stale" pipeline status.
   https://github.com/nexB/scancode.io/issues/130
 
+- Refactor the "execute" management command and add support for --async mode.
+  https://github.com/nexB/scancode.io/issues/130
+
 - Add a SCANCODEIO_REST_API_PAGE_SIZE setting to control the number of objects
   returned per page in the REST API.
   https://github.com/nexB/scancode.io/issues/328